+++ pricing/src/epilot/sdkconfiguration.py
@@ -24,9 +24,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.0.0'
-    sdk_version: str = '4.2.1'
-    gen_version: str = '2.283.1'
-    user_agent: str = 'speakeasy-sdk/python 4.2.1 2.283.1 1.0.0 epilot-pricing'
+    sdk_version: str = '4.2.2'
+    gen_version: str = '2.291.0'
+    user_agent: str = 'speakeasy-sdk/python 4.2.2 2.291.0 1.0.0 epilot-pricing'
     retry_config: RetryConfig = None
     _hooks: SDKHooks = None